Fix AdaptiveNutrition proteinRemaining copy to use Int grams type

+++ Fitness Coach/Application/StateBuilders/HealthIntelligence/HealthIntelligencePresentationCopy.swift
@@ -73,7 +73,7 @@ enum HealthIntelligencePresentationCopy {
                 FormaProductCopy.Today.HealthIntelligence.AdaptiveNutrition.extraWater(ml)
             }
 
-            static func proteinRemaining(_ grams: Double) -> String {
+            static func proteinRemaining(_ grams: Int) -> String {
                 FormaProductCopy.Today.HealthIntelligence.AdaptiveNutrition.proteinRemaining(grams)
             }
         }
